feat:@TgkCapture Moved the GitHub App-related routes from auth_controller to main_routes
